feat: add put method
