+++ package.json
@@ -29,8 +29,6 @@
     "devDependencies": {
         "@types/node": "^17.0.42",
         "@types/uuid": "^8.3.4",
-        "@typescript-eslint/eslint-plugin": "^5.9.0",
-        "@typescript-eslint/parser": "^5.9.0",
         "eslint": "^8.6.0",
         "eslint-config-airbnb-base": "^15.0.0",
         "eslint-config-airbnb-typescript": "^16.1.0",
