+++ src/interface.ts
@@ -1,5 +1,6 @@
 export interface IPerson {
   id?: string
   name: string,
-  age: number
+  age: number,
+  hobby: string[]
 }
