+++ src/server.ts
@@ -1,6 +1,9 @@
 import { createServer, IncomingMessage, ServerResponse } from 'http';
 import dotenv from 'dotenv';
-import { getUser, getUsers, createUser } from './controllers/userController';
+import { updateUser } from './controllers/updateUser';
+import { getUsers } from './controllers/getUsers';
+import { getUser } from './controllers/getUser';
+import { createUser } from './controllers/createUser';
 
 dotenv.config();
 
@@ -13,6 +16,8 @@ const server = createServer((req: IncomingMessage, res: ServerResponse) => {
     getUser(res, id);
   } else if (req.url === '/api/users' && req.method === 'POST') {
     createUser(req, res);
+  } else if (req.url?.includes('/api/users') && id.length > 0 && req.method === 'PUT') {
+    updateUser(res, req, id);
   } else {
     res.writeHead(404, { 'Content-Type': 'application/json' });
     res.end(JSON.stringify({ message: 'Route Not Found' }));
